Temporary Commit

+++ models/pizza.py
@@ -5,7 +5,7 @@ class Pizza(db.Model):
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     name = db.Column(db.String(80), unique=True, nullable=False)
     vegetarian = db.Column(db.Boolean, nullable=False)
-    price = db.Column(db.Integer, nullable=False)
+    price = db.Column(db.Float, nullable=False)
     ingredients = db.relationship('Ingredient', backref='pizza')
 
     @staticmethod
